添加代码

+++ src/main/java/com/gis09/fsm/common/config/FSMConfig.java
@@ -10,7 +10,7 @@ public class FSMConfig {
 	public static final long FSM_RECONNECT_INTERVAL = 60;// 重连间隔
 	public static final long FSM_RECONNECT_NUM = 5;// 默认的重试次数 重试五次放弃
 	public static final String FSM_DEFAULT_SERVER_HOST="127.0.0.1"; //默认绑定的地址
-	public static final long FSM_DEFAULT_PORT=52018;//默认的启动端口
+	public static final int FSM_DEFAULT_PORT=52018;//默认的启动端口
 	private long fsm_heart_interval=FSM_HEART_INTERVAL;
 	
 	private long fsm_reconnect_interval=FSM_RECONNECT_INTERVAL;
@@ -19,14 +19,14 @@ public class FSMConfig {
 	
 	private String fsm_server_host=FSM_DEFAULT_SERVER_HOST;
 	
-	private long fsm_port=FSM_DEFAULT_PORT;
+	private int fsm_port=FSM_DEFAULT_PORT;
 	public long getFsm_heart_interval() {
 		return fsm_heart_interval;
 	}
 	
 	
 	public FSMConfig(long fsm_heart_interval, long fsm_reconnect_interval,
-			long fsm_reconnect_num, String fsm_server_host, long fsm_port) {
+			long fsm_reconnect_num, String fsm_server_host, int fsm_port) {
 		super();
 		this.fsm_heart_interval = fsm_heart_interval;
 		this.fsm_reconnect_interval = fsm_reconnect_interval;
@@ -35,7 +35,7 @@ public class FSMConfig {
 		this.fsm_port = fsm_port;
 	}
 	
-	public FSMConfig(String fsm_server_host, long fsm_port) {
+	public FSMConfig(String fsm_server_host, int fsm_port) {
 		super();
 		this.fsm_server_host = fsm_server_host;
 		this.fsm_port = fsm_port;
@@ -74,11 +74,11 @@ public class FSMConfig {
 		this.fsm_server_host = fsm_server_host;
 	}
 
-	public long getFsm_port() {
+	public int getFsm_port() {
 		return fsm_port;
 	}
 
-	public void setFsm_port(long fsm_port) {
+	public void setFsm_port(int fsm_port) {
 		this.fsm_port = fsm_port;
 	}
